backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]>=2.0.48
alembic==1.14.1
asyncpg==0.30.0
python-dotenv==1.0.1
sendgrid==6.11.0
pyjwt==2.10.1
bcrypt>=4.0
python-multipart==0.0.20
email-validator>=2.0
pydantic-settings==2.7.1
cryptography>=42.0
qrcode>=7.4
